+++ sdk/individualAccountAttachment/log/index.js
@@ -1,6 +1,5 @@
 const log = require('./log.js');
 
-exports.Log = log.Log;
 exports.get = log.get;
 exports.query = log.query;
 exports.page = log.page;